+++ lib/src/controls/better_player_cupertino_controls.dart
@@ -628,6 +628,9 @@ class _BetterPlayerCupertinoControlsState
           isLoading(_controller.value)) {
         setState(() {
           _latestValue = _controller.value;
+          if (isVideoFinished(_latestValue)) {
+            _hideStuff = false;
+          }
         });
       }
     }