+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.31
+* Added showPlaceholderUntilPlay in BetterPlayerConfiguration
+* Fixed exception event not being triggered
+* Fixed controls not displaying on video finished
+
 ## 0.0.30
 * Fixed issue when full screen was triggered twice if autoPlay and fullScreenByDefault were enabled
 * Removed flutter_widgets, since it's not maintained anymore. Added instead visibility_detector package (by https://github.com/espresso3389)